+++ src/app/app.module.ts
@@ -12,6 +12,7 @@ import {FileFormComponent} from './components/file-form/file-form.component';
     ],
     imports: [
         BrowserModule,
+        FileInputAccessorModule,
         FileInputAccessorModule
     ],
     providers: [],